[MetaSchedule][BugFix] Fix broken integration tests
